Thursday late nigth - exc 1,2,3,4 and 5 complete :)

+++ java-bootcamp-bahia-t1/src/Item.java
@@ -2,7 +2,7 @@
 
 public interface Item {
 
-	public int getPrice();
+	public float getPrice();
 
 	public String getName();
 
@@ -12,7 +12,7 @@ public interface Item {
 
 	public void setCode(String code);
 
-	public void setPrice(int price);
+	public void setPrice(float price);
 
 	public String toString();
